+++ frontends/PyCDE/src/pycde/module.py
@@ -2,7 +2,7 @@
 #  See https://llvm.org/LICENSE.txt for license information.
 #  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from .hacks import attribute_to_var
+from __future__ import annotations
 
 from circt import support
 from circt.dialects import hw
@@ -261,7 +261,7 @@ class _Generate:
 
     # Assemble the parameters.
     self.params = {
-        nattr.name: attribute_to_var(nattr.attr)
+        nattr.name: support.attribute_to_var(nattr.attr)
         for nattr in mlir.ir.DictAttr(op.attributes["parameters"])
     }
